+++ test/hoomd_script/test_analyze_msd.py
@@ -35,8 +35,8 @@ class analyze_msd_tests (unittest.TestCase):
     
     def tearDown(self):
         init.reset();
-        os.remove("test.log");
-
+        if comm.get_rank() == 0:
+            os.remove("test.log");
 
 if __name__ == '__main__':
     unittest.main(argv = ['test.py', '-v'])
